+++ src/App.js
@@ -2,7 +2,7 @@ import "./App.css";
 import Home from "./pages/Home";
 import About from "./pages/About";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Footer from "./components/Footer";
+import Navbar from "./components/Navbar";
 import Contact from "./pages/Contact";
 import PageNotFound from "./pages/PageNotFound";
 import { Container } from "react-bootstrap";
@@ -11,7 +11,7 @@ function App() {
   return (
     <div className="App">
       <Router>
-        <Footer />
+        <Navbar />
         <Switch>
           <Route exact path={["/home", "/", ""]}>
             <Home />
